+++ naslib/predictors/ensemble.py
@@ -10,7 +10,6 @@ from naslib.predictors.gcn import GCNPredictor
 from naslib.predictors.bonas import BonasPredictor
 from naslib.predictors.bnn import DNGOPredictor, BOHAMIANN, BayesianLinearRegression
 from naslib.predictors.seminas import SemiNASPredictor
-from naslib.predictors.tpe import TreeParserEstimator
 from naslib.predictors.gp import (
     GPPredictor,
     SparseGPPredictor,
